feat(three-choice-quiz)intoroduced components  but my code still doesn't work
