--- src/routes/three-choice-quiz-practice/Answer.svelte
@@ -1,5 +1,0 @@
-<script>
-  export let answerText;
-  export let checkAnswer
-</script>
-
