+++ src/routes/three-choice-quiz-practice/question.js
@@ -1,0 +1,24 @@
+export const quizzes = [
+  {
+    question: "①出身地は？",
+    answers: ["愛知県", "高知県", "山口県"],//[]で囲むのが、配列のリテラル表現
+    correctAnswer: 0,
+  },
+  {
+    question: "②好きな動物は？",
+    answers: ["Tigers", "Doragons", "Giants"],
+    correctAnswer: 1,
+  },
+  {
+    question: "③好きな音楽は？",
+    answers: ["Country", "Classic", "Heavy Metal"],
+    correctAnswer: 0,
+  },
+  {
+    question: "④好きなアニメは？",
+    answers: ["マジンガーZ", "未来少年コナン", "アルプスの少女ハイジ"],
+    correctAnswer: 1,
+  },
+];
+
+
